feat: 기본 기능 구현

+++ src/com/iKeeper/Day1/Calculate.java
@@ -3,22 +3,19 @@ package com.iKeeper.Day1;
 
 
 public class Calculate {
-
-
-
-    public int add(int num1, int num2){
+    public float add(float num1, float num2){
         return num1+num2;
     }
 
-    public int sub(int num1, int num2){
+    public float sub(float num1, float num2){
         return num1-num2;
     }
 
-    public int mul(int num1, int num2){
+    public float mul(float num1, float num2){
         return num1*num2;
     }
 
-    public int div(int num1, int num2){
+    public float div(float num1, float num2){
         return num1/num2;
     }
 }
